# PctBackend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PCT UVCI — API REST",
        "description": "API de gestion des activités pédagogiques et volumes horaires — UVCI",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Serveur local"
        }
    ],
    "paths": {
...
        "/parametres": {
            "get": {
                "tags": [
                    "Paramètres"
                ],
                "summary": "Liste des coefficients VHN",
                "operationId": "e8a31bfd9325c4e8c92d8259776501b5",
                "responses": {
                    "200": {
                        "description": "Paramètres"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Paramètres"
                ],
                "summary": "Créer un paramètre",
                "operationId": "dc39c06e19410d50a135062bdaacdf98",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type_operation",
                                    "niveau_complexite",
                                    "coefficient_vhn"
                                ],
                                "properties": {
                                    "type_operation": {
                                        "type": "string",
                                        "enum": [
                                            "creation",
                                            "mise_a_jour"
                                        ]
                                    },
                                    "niveau_complexite": {
                                        "type": "string",
                                        "enum": [
                                            "simple",
                                            "intermediaire",
                                            "complexe"
                                        ]
                                    },
                                    "coefficient_vhn": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 1.5
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Créé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/parametres/{id}": {
            "put": {
                "tags": [
                    "Paramètres"
                ],
                "summary": "Modifier",
                "operationId": "ff0cb4dcd7a8717f4e5a6709358dbccb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "coefficient_vhn": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Modifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Paramètres"
                ],
                "summary": "Supprimer",
                "operationId": "641c004c7a113a5fa6e68f8042754a2d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Supprimé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Token obtenu via POST /login. Format : Bearer {token}",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}